# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Parbati Enterprises Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses with filters (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: projectId
          in: query
          schema:
            type: integer
        - name: particularId
          in: query
          schema:
            type: integer
        - name: paymentStatusId
          in: query
          schema:
            type: integer
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: voucherType
          in: query
          schema:
            type: string
            enum: [payment, receive]
        - name: voucherNumber
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Expense list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseListResponse"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense detail (includes documents)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense detail with documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense (creates new version) (expense_entry+)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseRequest"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherDetailResponse"
        "409":
          description: Expense is approved or finalized — cannot edit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Soft-delete expense (admin+)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /expenses/{id}/approve:
    patch:
      operationId: approveExpense
      tags: [expenses]
      summary: Approve an expense (accounts+). Blocks further edits.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved expense detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}/finalize:
    patch:
      operationId: finalizeExpense
      tags: [expenses]
      summary: Finalize an expense (accounts+). Requires prior approval. Permanently locks the record.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Finalized expense detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Not yet approved, or already finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}/history:
    get:
      operationId: getExpenseHistory
      tags: [expenses]
      summary: Get all versions of an expense (requires can_view_history flag)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExpenseVersion"
                required: [data]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    VoucherDetail:
      type: object
      properties:
        expenseId:
          type: integer
        voucherType:
          type: string
          enum: [payment, receive]
        projectId:
          type: integer
        projectName:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        createdBy:
          type: integer
        approvedBy:
          type: integer
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        finalizedBy:
          type: integer
          nullable: true
        finalizedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        versionId:
          type: integer
        versionNo:
          type: integer
        voucherNumber:
          type: string
        expenseDate:
          type: string
          format: date
        particularId:
          type: integer
        particularName:
          type: string
        description:
          type: string
          nullable: true
        transactionDetails:
          type: string
          nullable: true
        uomId:
          type: integer
        uomName:
          type: string
        quantity:
          type: string
        pricePerUnit:
          type: string
        amount:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        paymentStatusId:
          type: integer
        paymentStatusName:
          type: string
      required:
        - expenseId
        - voucherType
        - projectId
        - projectName
        - vendorId
        - vendorName
        - createdBy
        - createdAt
        - updatedAt
        - versionId
        - versionNo
        - voucherNumber
        - expenseDate
        - particularId
        - particularName
        - uomId
        - uomName
        - quantity
        - pricePerUnit
        - amount
        - paymentStatusId
        - paymentStatusName
    VoucherDetailResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/VoucherDetail"
      required: [data]
    ExpenseDetailResponse:
      type: object
      properties:
        data:
          allOf:
            - $ref: "#/components/schemas/VoucherDetail"
            - type: object
              properties:
                documents:
                  type: array
                  items:
                    $ref: "#/components/schemas/Document"
              required: [documents]
      required: [data]
    ExpenseListItem:
      type: object
      properties:
        expenseId:
          type: integer
        voucherType:
          type: string
          enum: [payment, receive]
        projectName:
          type: string
        vendorName:
          type: string
        voucherNumber:
          type: string
        expenseDate:
          type: string
          format: date
        particularName:
          type: string
        uomName:
          type: string
        quantity:
          type: string
        pricePerUnit:
          type: string
        amount:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        paymentStatusName:
          type: string
        paymentStatusId:
          type: integer
        versionId:
          type: integer
        versionNo:
          type: integer
        approvedAt:
          type: string
          format: date-time
          nullable: true
        finalizedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - expenseId
        - voucherType
        - projectName
        - vendorName
        - voucherNumber
        - expenseDate
        - particularName
        - uomName
        - quantity
        - pricePerUnit
        - amount
        - paymentStatusName
        - paymentStatusId
        - versionId
        - versionNo
    ExpenseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseListItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
...
    UpdateExpenseRequest:
      type: object
      properties:
        expenseDate:
          type: string
          format: date
        particularId:
          type: integer
        description:
          type: string
        transactionDetails:
          type: string
        uomId:
          type: integer
        quantity:
          type: number
        pricePerUnit:
          type: number
        invoiceNumber:
          type: string
        paymentStatusId:
          type: integer
    ExpenseVersion:
      type: object
      properties:
        id:
          type: integer
        versionNo:
          type: integer
        voucherNumber:
          type: string
        expenseDate:
          type: string
          format: date
        particularName:
          type: string
        uomName:
          type: string
        quantity:
          type: string
        pricePerUnit:
          type: string
        amount:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        paymentStatusName:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - versionNo
        - voucherNumber
        - expenseDate
        - particularName
        - uomName
        - quantity
        - pricePerUnit
        - amount
        - paymentStatusName
    Document:
      type: object
      properties:
        id:
          type: integer
        expenseVersionId:
          type: integer
        voucherNumber:
          type: string
        originalName:
          type: string
        storedName:
          type: string
        mimeType:
          type: string
          nullable: true
        fileSize:
          type: integer
        publicUrl:
          type: string
          nullable: true
        fileOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - expenseVersionId
        - voucherNumber
        - originalName
        - storedName
        - fileSize
        - fileOrder
...